requirements.txt
PyYAML==6.0.1
aiogram==3.6.0
annotated-types==0.6.0
anyio==3.6.2
click==8.1.7
fastapi==0.109.1
h11==0.14.0
httptools==0.6.1
idna==3.6
pip==24.1.2
pydantic==2.6.0
pydantic-core==2.16.1
python-dotenv==1.0.1
setuptools==68.2.0
sniffio==1.3.0
starlette==0.35.1
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.41.2
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
passlib==1.7.4
makefun==1.15.2
python-multipart==0.0.9
pyjwt==2.8.0
pycparser==2.21
dnspython==2.5.0
bcrypt==4.1.2
email-validator==2.0.0.post2
cffi==1.16.0
cryptography==42.0.2
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
pydantic-settings==2.1.0
seatable-api==2.6.11
fastapi-pagination==0.12.24
fastapi-filter==2.0.0
APScheduler==3.10.4
pytest==7.2.1
pytest-asyncio==0.20.3
aioredis==2.0.1
aiosqlite==0.20.0

# Celery
amqp==5.3.1
billiard==4.2.1
celery==5.4.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.4.2
prompt-toolkit==3.0.48
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
requests==2.32.3
psycopg2-binary==2.9.10
redis==5.2.0

# OpenAI
distro==1.9.0
httpcore==1.0.7
httpx==0.27.2
jiter==0.7.1
openai==1.54.4
tqdm==4.67.0
typing-extensions==4.12.2
